requirements.txt
Flask
openai
python-dotenv
pandas
flask-cors